Fix: React implementation, lint, and build.
