--- React/orig_package.json
@@ -1,55 +1,0 @@
-{
-  "name": "react",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
-    "test": "vitest",
-    "test:watch": "vitest --watch",
-    "lint-js": "eslint .",
-    "lint-css": "stylelint src/**/*.css --allow-empty-input",
-    "lint": "npm-run-all -p -c lint-js lint-css",
-    "preview": "vite preview"
-  },
-  
-  "dependencies": {
-    "devextreme": "^24.2.3",
-    "devextreme-react": "^24.2.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-   "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^14.1.2",
-    "@testing-library/user-event": "^14.4.3",
-    "@types/react": "^18.2.47",
-    "@types/react-dom": "^18.2.17",
-    "@vitejs/plugin-react": "^4.4.1",
-    "@vitest/coverage-v8": "^1.5.0",
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "eslint": "^8.35.0",
-    "eslint-config-devextreme": "^1.1.4",
-    "eslint-plugin-no-only-tests": "2.6.0",
-    "eslint-plugin-react": "7.31.10",
-    "eslint-plugin-react-hooks": "^5.2.0",
-    "eslint-plugin-react-perf": "^3.3.1",
-    "globals": "^16.0.0",
-    "jsdom": "^24.0.0",
-    "npm-run-all": "^4.1.5",
-    "stylelint": "^15.6.1",
-    "stylelint-config-standard": "^33.0.0",
-    "typescript": "~5.8.2",
-    "typescript-eslint": "^8.18.2",
-    "vite": "^6.3.5",
-    "ts-node": "10.9.2",
-    "vitest": "^1.5.0"
-  }
-}
