--- React/src/orig_App.css
@@ -1,4 +1,0 @@
-.main {
-  margin: 50px;
-  width: 90vh;
-}
